third_party/blink/web_tests/platform/linux-chrome/external/wpt/storage-access-api/storage-access-beyond-cookies.none.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify StorageAccessAPIBeyondCookies for None
  assert_equals: Storage Access API should not allow access for empty requests. expected (string) "HasAccess for none" but got (object) object "[object Object]"
Harness: the test ran to completion.
